requirements.txt
# Jarvis Discord Voice Assistant Dependencies

# Discord Bot
discord.py>=2.0.0
PyNaCl>=1.5.0

# Web Server
fastapi>=0.109.0
uvicorn>=0.27.0

# STT (Speech-to-Text)
openai-whisper>=20231117

# LLM (Ollama)
ollama>=0.4.0

# TTS (Text-to-Speech)
edge-tts>=6.1.9

# Music Playback
yt-dlp>=2024.1.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
